requirements.txt
langchain
langchain-openai
langchain-community
langchain-tavily
yfinance
pandas
streamlit
python-dotenv
matplotlib